Update README.md
Feeling more positive and in the right directions. Bits and pcs. coming back. I can only think of a fix for the mistake I made today unless it was a beta effect of directional path order of installments or my OS. Win.7 64 bit Ultimate. Waiting on an answer or decision. Unless this was a Beta test with or without a fix. Or still working on the updates. I can delete, clean and reinstall or find the error with your help instead of doing it myself. could just be a missing file. Be talking with you soon master and Read- Me, and ".... - Edit_ Pleasure getting this far and meeting you. It is just the beginning for me. Nice day.